Scope FOV dynamic adjust
Scope FOV dynamic adjust
Seperated hand compoennt script
added external config file.
