+++ artifact/scripts/CONFIG.lua
@@ -1,0 +1,7 @@
+	 isRhand = true
+	 HapticFeedback = true
+	 isLeftHandModeTriggerSwitchOnly = true
+	
+	 SwapLShoulderLThumb =true --sprinting switch with l shoulder
+	 SitMode=false
+
